+++ Blog.Domain/Services/PostService.cs
@@ -18,9 +18,9 @@ namespace Blog.Domain.Services
             _postRepository = postRepository;
         }
 
-        public async Task<IEnumerable<PostDao>> GetPosts()
+        public async Task<IEnumerable<PostDao>> GetPosts(int userId = 0)
         {
-            var posts = await _postRepository.GetPosts();
+            var posts = await _postRepository.GetPosts(userId);
             return posts.Select(p => new PostDao(p)).ToList();
         }
 
@@ -51,22 +51,28 @@ namespace Blog.Domain.Services
             await _postRepository.InsertPost(post);
         }
 
-        public async Task<bool> UpdatePost(Post post)
+        public async Task<bool> UpdatePost(Post post, bool isAdmin, int currentUserId)
         {
             var postFound = await _postRepository.GetPost(post.PostId);
 
             if (postFound == null)
                 throw new Exception("Post does not exist");
 
+            if (!isAdmin && postFound.User.UserId != currentUserId)
+                throw new Exception("User does not have permission to perform this action");
+
             return await _postRepository.UpdatePost(post);
         }
 
-        public async Task<bool> DeletePost(int id)
+        public async Task<bool> DeletePost(int id, bool isAdmin, int currentUserId)
         {
             var postFound = await _postRepository.GetPost(id);
 
             if (postFound == null)
                 throw new Exception("Post does not exist");
+
+            if (!isAdmin && postFound.User.UserId != currentUserId)
+                throw new Exception("User does not have permission to perform this action");
 
             return await _postRepository.DeletePost(id);
         }
